Bug fixes, attempts at adding TBC spells
I made these changes a while ago but I no longer have time nor motivation to continue developing this.

Partly it's because I don't have motivation for it (I have multiple other projects that I'm working on that take precedence for me), and
there are already (better) addons out there that does what FieldGuide does.

Partly it's because I realized I'd have to rewrite a lot of the code and I don't have time to get into how it works again.
(Really wish I had commented better and written cleaner code when I wrote this 2-3 years ago.)

Pushing these changes so that if someone else maybe wants to try to update it for TBC, it'll perhaps help a bit.